insert user

+++ src/app/pojos/User.java
@@ -1,5 +1,5 @@
 package app.pojos;
-// Generated Dec 18, 2020 3:57:46 PM by Hibernate Tools 4.3.1
+// Generated Dec 18, 2020 5:10:53 PM by Hibernate Tools 4.3.1
 
 
 
@@ -14,21 +14,21 @@ public class User  implements java.io.Serializable {
      private String nickname;
      private String email;
      private String avatar;
-     private byte isDeleted;
+     private int isDeleted;
      private String password;
 
     public User() {
     }
 
 	
-    public User(String username, String nickname, String email, byte isDeleted, String password) {
+    public User(String username, String nickname, String email, int isDeleted, String password) {
         this.username = username;
         this.nickname = nickname;
         this.email = email;
         this.isDeleted = isDeleted;
         this.password = password;
     }
-    public User(String username, String nickname, String email, String avatar, byte isDeleted, String password) {
+    public User(String username, String nickname, String email, String avatar, int isDeleted, String password) {
        this.username = username;
        this.nickname = nickname;
        this.email = email;
@@ -72,11 +72,11 @@ public class User  implements java.io.Serializable {
     public void setAvatar(String avatar) {
         this.avatar = avatar;
     }
-    public byte getIsDeleted() {
+    public int getIsDeleted() {
         return this.isDeleted;
     }
     
-    public void setIsDeleted(byte isDeleted) {
+    public void setIsDeleted(int isDeleted) {
         this.isDeleted = isDeleted;
     }
     public String getPassword() {
